pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qimg"
version = "0.1.0"
description = "Query Images - local-first image search CLI"
readme = "README.md"
requires-python = ">=3.13,<3.14"
authors = [{ name = "qimg contributors" }]
license = { text = "MIT" }
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.0",
  "PyYAML>=6.0.1",
  "Pillow>=10.3.0",
  "ImageHash>=4.3.1",
  "numpy>=2.0.0",
  "pydantic>=2.8.2",
]

[project.optional-dependencies]
hnsw = ["hnswlib>=0.8.0"]
encoder = ["torch>=2.4.0", "transformers>=4.48.0", "huggingface_hub>=0.30.0", "safetensors>=0.5.0", "sentencepiece>=0.2.0", "protobuf>=5.27.0"]
openclip = ["torch>=2.4.0"]
ocr = ["pytesseract>=0.3.10"]
exif = ["piexif>=1.1.3"]
test = ["pytest>=8.2.0"]
finetune = ["torch>=2.4.0", "transformers>=4.48.0", "huggingface_hub>=0.30.0", "safetensors>=0.5.0", "sentencepiece>=0.2.0", "protobuf>=5.27.0"]

[project.scripts]
qimg = "qimg.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
